refactor: split handlers/boot.ts into four files by concern
`handlers/boot.ts` had grown to 1661 lines — the second-largest file
in the codebase after schema.ts. It mixed payload types, derivation
logic, text rendering, and the dispatcher. Split into:

- bootTypes.ts        (250 lines) — payload shapes (BootPayload,
                                    BootSuggestedNext, BootBroadcastPendingResponse,
                                    BootActiveOverlap, ActionableKind,
                                    ActionableTransition). Leaf module —
                                    type-only, no runtime code.
- bootActionable.ts   (650 lines) — all derivation logic:
                                    actionableTransitions,
                                    deriveBootSuggestedNext,
                                    derivePendingBroadcastResponse,
                                    deriveVerbsAvailableNow,
                                    computeActiveOverlappingTargets,
                                    computeLastAuthoredWriteAt,
                                    stampActorResolved,
                                    isBroadcastPendingResponse +
                                    ACTIONABLE_PRIORITY,
                                    REVIEWED_AUTHORED_ACTIONABLE_CAP,
                                    ALWAYS_READABLE_VERBS.
- bootRender.ts       (320 lines) — text-mode renderer +
                                    PASSAGE_ORIENTATION_REGISTRY +
                                    collectCrossPassage.
- boot.ts            (370 lines) — bootCmd orchestrator + KNOWN_FLAGS
                                    + re-exports of the externally-
                                    consumed helpers (deriveBootSuggestedNext,
                                    BootSuggestedNext, computeLastAuthoredWriteAt).

Total: 1590 lines (was 1661 → -71 by removing duplicate JSDoc; new
module headers compensate roughly evenly).

Each file now has one concern. boot.ts itself shrank from 1661 to
**370 lines** (-78%). No behavior change — every handler's output
and lifecycle remain byte-identical. Full suite passes unchanged
(1677/1677).

Leaf-type-module layout (bootTypes.ts has no runtime dependencies)
was chosen specifically to avoid the circular-import shape that
would otherwise emerge: bootCmd needs derivation that needs the
payload shape, but the payload shape can't live in boot.ts because
bootActionable imports types. The leaf isolation breaks the cycle.

External consumers (`suggest.ts`, `tests/interface/boot.test.ts`)
keep their original `import ... from './boot.js'` paths via
re-exports — no test or downstream module changes needed.
